pyproject.toml
[tool.poetry]
name = "space"
version = "0.1.0"
description = "Space AG 48h Challenge"
authors = ["Oscar Giraldo Castillo <oscar.gi.cast@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
psycopg2-binary = "^2.9.1"
dj-database-url = "^0.5.0"
django-model-utils = "^4.1.1"
environs = "^9.3.3"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
ipython = "^7.27.0"
pytest-django = "^4.4.0"
model-bakery = "^1.3.2"
factory-boy = "^3.2.0"
PyHamcrest = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
